+++ src/adt/AtomicArena.hh
@@ -13,10 +13,10 @@ struct AtomicArena : Allocator
     Arena arena; /* compose for 1 mutex instead of second mutex for realloc (or recursive mutex) */
 
     AtomicArena() = default;
-    AtomicArena(size_t blockSize) : arena(blockSize) { mtx_init(&this->mtxA, mtx_plain); }
+    AtomicArena(u32 blockSize) : arena(blockSize) { mtx_init(&this->mtxA, mtx_plain); }
 
     virtual void*
-    alloc(size_t memberCount, size_t size) override
+    alloc(u32 memberCount, u32 size) override
     {
         mtx_lock(&this->mtxA);
         auto rp = this->arena.alloc(memberCount, size);
@@ -34,7 +34,7 @@ struct AtomicArena : Allocator
     }
 
     virtual void*
-    realloc(void* p, size_t size) override
+    realloc(void* p, u32 size) override
     {
         mtx_lock(&this->mtxA);
         auto rp = this->arena.realloc(p, size);
